+++ test/Utils.hs
@@ -8,7 +8,7 @@ satTest :: D.SMTResult -> IO ()
 satTest D.SolverSat{} = return ()
 satTest e = error $ unwords ["Expected a SAT result but got", show e]
 
-vtest :: D.SMTResult -> M.Map String Integer -> IO ()
+vtest :: D.SMTResult -> M.Map String Float -> IO ()
 vtest result expectedVars = case result of
   D.SolverUnsat -> error "Expected SAT but got UNSAT"
   D.SolverFailed -> error "Expected SAT but the solver failed"
